feat!: Optional sources
Files and possibly other types of sources may not be present all the time.
Some sources may only be present in specific environments, and
fallback to a lower-priority source/default configuration is
an explicitly chosen strategy. To support this scenario, sources
are no longer required to provide a result. When `None` is returned,
the source is seen as not present and **ignored**. As not all sources
are expected to be optional, this is source-specific. Each source
manages optional presence on its own.